requirements.txt
pillow
pyyaml
requests
icalendar
pycaruna
gkeepapi
